// www/src/lib/db/database.types.ts
export type Database = {
	graphql_public: {
		Tables: {
			[_ in never]: never;
		};
		Views: {
			[_ in never]: never;
		};
		Functions: {
			graphql: {
				Args: {
					operationName?: string;
					query?: string;
					variables?: Json;
					extensions?: Json;
				};
				Returns: Json;
			};
		};
		Enums: {
			[_ in never]: never;
		};
		CompositeTypes: {
			[_ in never]: never;
		};
	};
	public: {
		Tables: {
			computers: {
				Row: {
					created_at: string;
					is_added: boolean;
					is_allowed: boolean;
					one_time_key: string | null;
					org_unit_uuid: string;
					tenant_uuid: string;
					uuid: string;
				};
				Insert: {
					created_at?: string;
					is_added?: boolean;
					is_allowed?: boolean;
					one_time_key?: string | null;
					org_unit_uuid?: string;
					tenant_uuid?: string;
					uuid?: string;
				};
				Update: {
					created_at?: string;
					is_added?: boolean;
					is_allowed?: boolean;
					one_time_key?: string | null;
					org_unit_uuid?: string;
					tenant_uuid?: string;
					uuid?: string;
				};
				Relationships: [
					{
						foreignKeyName: 'public_computer_org_unit_uuid_fkey';
						columns: ['org_unit_uuid'];
						isOneToOne: false;
						referencedRelation: 'org_units';
						referencedColumns: ['uuid'];
					},
					{
						foreignKeyName: 'public_computer_tenant_uuid_fkey';
						columns: ['tenant_uuid'];
						isOneToOne: false;
						referencedRelation: 'tenants';
						referencedColumns: ['uuid'];
					}
				];
			};
			org_units: {
				Row: {
					color: string | null;
					created_at: string;
					icon_id: number;
					name: string;
					tenant_uuid: string;
					uuid: string;
				};
				Insert: {
					color?: string | null;
					created_at?: string;
					icon_id?: number;
					name?: string;
					tenant_uuid?: string;
					uuid?: string;
				};
				Update: {
					color?: string | null;
					created_at?: string;
					icon_id?: number;
					name?: string;
					tenant_uuid?: string;
					uuid?: string;
				};
				Relationships: [
					{
						foreignKeyName: 'public_org_unit_tenant_uuid_fkey';
						columns: ['tenant_uuid'];
						isOneToOne: false;
						referencedRelation: 'tenants';
						referencedColumns: ['uuid'];
					}
				];
			};
// ...
		};
		Views: {
			[_ in never]: never;
		};
		Functions: {
			[_ in never]: never;
		};
		Enums: {
			machine_type: 'LXC' | 'VM' | 'Physical' | 'Unknown';
		};
		CompositeTypes: {
			[_ in never]: never;
		};
	};
// ...
};